ajustes com base no guides
